test IR functional and make sure Mcu2120Init can be deleted and PB6, PB7 and PB15 can be used)

+++ 0088-V-U-038-A/DRV/autonetAPIs.c
@@ -423,8 +423,8 @@ uint8_t get_gps(uint8_t* Lat_deg, uint8_t* Lat_min, uint8_t* Lat_sec, uint8_t* L
 uint8_t get_velocity(int* speed){
 		
 }
-
-uint8_t IR_read(uint8_t IR_BufferRx, uint8_t length, uint8_t index){
+/*
+uint8_t IR_read(uint8_t IR_BufferRx, unsigned short length, uint8_t index){
 	// TODO: to ensure the correctness
 	Mcp2120Proc(&IR_BufferRx, &length, index);
 	Delay(10);
@@ -435,6 +435,7 @@ uint8_t IR_write(uint8_t IR_BufferTx, uint8_t length, uint8_t index){
 	Mcp2120Tx(&IR_BufferTx, length , index);
 	Delay(10);
 }
+*/
 
 // --------------  Matnetometer functions by chih-wei -----------------
 void Mag_Error_Handle (short *pX, short *pY,short *pZ, short *max_x, short *min_x , short *max_y, short *min_y, short *max_z, short *min_z)
